added inode structure to project and changed some return types

+++ include/afs/disk.h
@@ -1,23 +1,24 @@
 #pragma once
 
 #include <cstdlib>
+#include <cstdint>
 
 class Disk
 {
 private:
     int fd;
     unsigned char* m_fileMap;
-    size_t m_blockSize;
-    size_t m_nblocks;
+    uint32_t m_blockSize;
+    uint32_t m_nblocks;
 
     void createDiskFile(const char* filePath);
 
 public:
-    Disk(const char* filePath, const size_t blockSize = 4096, const size_t nblocks = 4096);
+    Disk(const char* filePath, const uint32_t blockSize = 4096, const uint32_t nblocks = 4096);
     ~Disk();
 
-    size_t getBlockSize() { return m_blockSize; }
-    size_t getBlocksAmount() { return m_nblocks; }
+    uint32_t getBlockSize() { return m_blockSize; }
+    uint32_t getBlocksAmount() { return m_nblocks; }
     size_t getDiskSize() { return m_blockSize * m_nblocks; }
 
     void read(unsigned long addr, int size, char* ans);
